// types.ts
export type ViewState = 
  | 'auth'
  | 'home' 
  | 'analytics' 
  | 'teams' 
  | 'create-team'
  | 'profile' 
  | 'add-training' 
  | 'add-training-skiing'
  | 'activity-select' 
  | 'team-details' 
  | 'exercise-details' 
  | 'jump-details'
  | 'body-metrics'
  | 'all-body-metrics'
  | 'all-sessions'
  | 'coach-event-details'; // New View

export type UnitSystem = 'metric' | 'imperial';
// ...
export interface PRLog {
  id: string;
  exerciseId: string;
  date: string;
  weight: number; // Stored in kg
  note?: string;
}